move ogcapi-p to services module
