+++ eventkit_cloud/jobs/helpers.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 import os
