+++ eventkit_cloud/utils/mapproxy.py
@@ -8,18 +8,14 @@ from multiprocessing import Process
 from multiprocessing.dummy import DummyProcess
 from typing import Any, Dict, Tuple, cast
 
-import mapproxy
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.contrib.gis.geos import GEOSGeometry
 from django.core.cache import cache
 from django.db import connections
-from mapproxy.config.config import load_config, load_default_config
-from mapproxy.config.loader import ConfigurationError, ProxyConfiguration, validate_references
-from mapproxy.grid import tile_grid
-from mapproxy.wsgiapp import MapProxyApp
 from webtest import TestApp
 
+import mapproxy
 from eventkit_cloud.core.helpers import get_cached_model
 from eventkit_cloud.jobs.helpers import get_valid_regional_justification
 from eventkit_cloud.tasks import get_cache_value
@@ -34,6 +30,10 @@ from eventkit_cloud.utils.geopackage import (
     set_gpkg_contents_bounds,
 )
 from eventkit_cloud.utils.stats.eta_estimator import ETA
+from mapproxy.config.config import load_config, load_default_config
+from mapproxy.config.loader import ConfigurationError, ProxyConfiguration, validate_references
+from mapproxy.grid import tile_grid
+from mapproxy.wsgiapp import MapProxyApp
 
 # Mapproxy uses processes by default, but we can run child processes in demonized process, so we use
 # a dummy process which relies on threads.  This fixes a bunch of deadlock issues which happen when using billiard.
