+++ eventkit_cloud/utils/stats/generator.py
@@ -14,14 +14,14 @@ from django import db
 from django.core.cache import cache
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models.query import QuerySet
-from mapproxy import grid as mapproxy_grid
-from mapproxy import srs as mapproxy_srs
 
 from eventkit_cloud.jobs.models import DataProvider
 from eventkit_cloud.tasks.enumerations import TaskState
 from eventkit_cloud.tasks.models import DataProviderTaskRecord, ExportRun, ExportTaskRecord
 from eventkit_cloud.utils.client import parse_duration
 from eventkit_cloud.utils.stats.geomutils import get_area_bbox, get_bbox_intersect, get_geometry_description
+from mapproxy import grid as mapproxy_grid
+from mapproxy import srs as mapproxy_srs
 
 logger = logging.getLogger(__name__)
 _dbg_geom_cache_misses = 0
